Fixed numbering of steps
There was a formatting issue that caused step "2" to render as "1". It was indeed "2" in the code, but because the back-tick blocks were indented at the same level as the number 1, it was causing the second number to count as the start of a new list, and thus receive no. "1". I fixed that, and now it indeed says both step "1", and "2" for build instructions on the Mac.